+++ OrcoWeb/src/main/java/com/criterya/orco/spring/services/TimesService.java
@@ -8,6 +8,6 @@ public interface TimesService {
 
 	public List<Times> getAllTimes();
 	public Times getTime(Integer id);
-	public Long saveOrUpdate(Times time);
+	public Integer saveOrUpdate(Times time);
 	public void delete(Times time);
 }
